attached_assets/requirements_1762982955526.txt
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
cymem==2.0.7
decorator==4.4.2
ffmpeg-python==0.2.0
filelock==3.12.0
future==0.18.3
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-speech==2.20.0
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.1
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
imageio==2.30.0
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
langcodes==3.3.0
lit==16.0.5.post0
llvmlite==0.39.1
MarkupSafe==2.1.3
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
packaging==23.1
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydub==0.25.1
pyphen==0.14.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-syllables==3.0.2
srsly==2.4.6
sympy==1.12
thinc==8.1.10
tiktoken==0.3.1
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.3
urllib3==1.26.16
wasabi==1.1.1
numba==0.56.4
numpy==1.23.5
streamlit>=1.28.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.0
